Move SmartTwinSigned to Api middleware namespace and add EventType enum
- Move SmartTwinSigned from Middleware/ to Middleware/Api/ for consistency
- Update bootstrap/app.php alias to new namespace
- Add App\Enums\SmartTwin\EventType with RESIDENT_SCAN_FINISHED and COACH_SCAN_FINISHED
