+++ bootstrap/app.php
@@ -74,7 +74,7 @@ return Application::configure(basePath: dirname(__DIR__))
 
         $middleware->alias([
             'access.cooperation' => \App\Http\Middleware\Api\AllowIfTokenCanAccessCooperation::class,
-            'smarttwin.signed' => \App\Http\Middleware\SmartTwinSigned::class,
+            'smarttwin.signed' => \App\Http\Middleware\Api\SmartTwinSigned::class,
             'checks-conditions-for-sub-steps' => \App\Http\Middleware\ChecksConditionsForSubSteps::class,
             'cooperation' => \App\Http\Middleware\CooperationMiddleware::class,
             'cooperation-has-scan' => \App\Http\Middleware\CooperationHasScan::class,
